[FIX] 백엔드 return 타입 변경에 따른 response 타입 변경, 이미 로그인이 되어있을 시 페이지 엔터 방지
백엔드 return 타입 변경에 따른 response 타입 변경, 이미 로그인이 되어있을 시 페이지 엔터 방지
